ARXIVNG-1389 more work on abs macro, adding in filters, patch stylesheet
