+++ arxiv/base/templates/base/testmacros.html
@@ -17,6 +17,7 @@
     primary_category,
     version=version,
     submission_history=submission_history,
-    comments=comments) }}
+    comments=comments,
+    doi=doi) }}
 
 {% endblock %}
